Style

+++ patchwork-events-entity/src/main/java/net/patchworkmc/mixin/event/entity/entity/MixinEntity.java
@@ -1,3 +1,22 @@
+/*
+ * Minecraft Forge, Patchwork Project
+ * Copyright (c) 2016-2020, 2019-2020
+ *
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation version 2.1
+ * of the License.
+ *
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this library; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
+ */
+
 package net.patchworkmc.mixin.event.entity.entity;
 
 import net.minecraftforge.common.MinecraftForge;
@@ -5,10 +24,8 @@ import net.minecraftforge.event.entity.EntityEvent;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Overwrite;
 import org.spongepowered.asm.mixin.Shadow;
-import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.Redirect;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 import net.minecraft.entity.Entity;
@@ -87,18 +104,18 @@ public abstract class MixinEntity {
 		this.dimensions = entityDimensions;
 		this.standingEyeHeight = sizeEvent.getNewEyeHeight();
 		// forge end
+
 		if (entityDimensions2.width < entityDimensions.width) {
-			double d = (double)entityDimensions2.width / 2.0D;
-			this.setBoundingBox(new Box(this.getX() - d, this.getY(), this.getZ() - d, this.getX() + d, this.getY() + (double)entityDimensions2.height, this.getZ() + d));
+			double d = (double) entityDimensions2.width / 2.0D;
+			this.setBoundingBox(new Box(this.getX() - d, this.getY(), this.getZ() - d, this.getX() + d, this.getY() + (double) entityDimensions2.height, this.getZ() + d));
 		} else {
 			Box box = this.getBoundingBox();
-			this.setBoundingBox(new Box(box.minX, box.minY, box.minZ, box.minX + (double)entityDimensions2.width, box.minY + (double)entityDimensions2.height, box.minZ + (double)entityDimensions2.width));
+			this.setBoundingBox(new Box(box.minX, box.minY, box.minZ, box.minX + (double) entityDimensions2.width, box.minY + (double) entityDimensions2.height, box.minZ + (double) entityDimensions2.width));
+
 			if (entityDimensions2.width > entityDimensions.width && !this.firstUpdate && !this.world.isClient) {
 				float f = entityDimensions.width - entityDimensions2.width;
-				this.move(MovementType.SELF, new Vec3d((double)f, 0.0D, (double)f));
+				this.move(MovementType.SELF, new Vec3d(f, 0.0D, f));
 			}
-
 		}
 	}
-
 }
